requirements.txt
py3-bencode==0.0.3